Remove equality check for arr[low] and arr[mid]
 if (arr[mid] < arr[start]) {
                end = mid - 1;
            } 

This condition checks if arr[mid] and arr[low] are equal . 
But elements are distinct according to the question. So, this condition will only be true when array size will be 2 . 
For example [7, 2] , but this case will already be handled by above if statements.
so, there is no need to check equality.