Update ASSIGNMENT_SOLUTIONS.md
Hi, my name is Mohammad Hussain student of computer engineering also followed this course and try to contribute to this. When I started learning to problem-solve with DSA then I make this repository with imp questions on each topic there is also solutions of your playlist assignment. So kindly accept my repository.