Create SelectionSort.java
Hey Kunal! The code for selection sort in your lecture was by taking the maximum element at that last index. I did the same but with the minimum element to the first index. Here, the code is basically small and easy to understand. If you liked it, please accept my PR so that others can see it.