Checking in the boundary conditions manually
First of all I was confused how the pivot could return -1 in the method findPivot for a long time but later in the following testcase,

nums = {3,5,1} and target = 3

Finding pivot will return -1 which implies that the array is not a rotated array. But in reality we can obviously see that this array is rotated 2 times. Technically it works when the (nums[mid]  > nums[mid  + 1] ) condition in findPivot method but it's not how we wrote the logic.

So I came up with the condition 

(nums[0] < nums[arr.length - 1]) 

as the condition to check if it's a non rotated array. But this will again in the boundary conditions i.e. if the pivot lies in the 0th index or nums.length - 1 index because of the conditions we specified in the pivot conditions.

I couldn't come up with any solution for the boundary conditions other than checking it manually from line 11 to 17. Could you please help me with this kunal.